refactor: improve structure of reference and system tokens
* add focusIndicator tokens
* add shadow overrides
* include mass in toRawSpring
* add contentColorFor helper
