+++ src/components/Checkbox/utils.ts
@@ -3,7 +3,7 @@ import type { ColorValue } from 'react-native';
 import { tokens } from '../../theme/tokens';
 import type { InternalTheme } from '../../types';
 
-const { stateOpacity } = tokens.md.ref;
+const stateOpacity = tokens.md.sys.state.opacity;
 
 const getCheckedColor = ({
   theme,
