+++ src/components/RadioButton/utils.ts
@@ -3,7 +3,7 @@ import type { ColorValue, GestureResponderEvent } from 'react-native';
 import { tokens } from '../../theme/tokens';
 import type { InternalTheme } from '../../types';
 
-const { stateOpacity } = tokens.md.ref;
+const stateOpacity = tokens.md.sys.state.opacity;
 
 export const handlePress = ({
   onPress,
