+++ src/components/__tests__/RadioButton/utils.test.tsx
@@ -2,7 +2,7 @@ import { getTheme } from '../../../core/theming';
 import { tokens } from '../../../theme/tokens';
 import { getSelectionControlIOSColor } from '../../RadioButton/utils';
 
-const { stateOpacity } = tokens.md.ref;
+const stateOpacity = tokens.md.sys.state.opacity;
 
 describe('getSelectionControlIOSColor - checked color', () => {
   it('should return correct disabled color, for theme version 3', () => {
