--- src/theme/types/utils.ts
@@ -1,9 +1,0 @@
-import type * as React from 'react';
-
-export type $Omit<T, K> = Pick<T, Exclude<keyof T, K>>;
-export type $RemoveChildren<T extends React.ComponentType<any>> = $Omit<
-  React.ComponentPropsWithoutRef<T>,
-  'children'
->;
-
-export type EllipsizeProp = 'head' | 'middle' | 'tail' | 'clip';
