+++ src/types.tsx
@@ -1,1 +1,11 @@
+import type * as React from 'react';
+
 export * from './theme/types';
+
+export type $Omit<T, K> = Pick<T, Exclude<keyof T, K>>;
+export type $RemoveChildren<T extends React.ComponentType<any>> = $Omit<
+  React.ComponentPropsWithoutRef<T>,
+  'children'
+>;
+
+export type EllipsizeProp = 'head' | 'middle' | 'tail' | 'clip';
